Verify all Shares have same HMAC and resulting Secret has valid HMAC.
This ties the generated secret back to the original and verifies a
perfect match every time. Now we can be assured that the
re-hydrated secret is in fact exactly the same as the secret
the shares were generated from even though we no longer have
access to it. Also may help prevent cheating.